Rename a few classes to match requirements
