+++ Console.ExerciseTracker.ananttuli/App/Program.cs
@@ -1,9 +1,9 @@
-﻿using App.Database;
+﻿using App.Controllers;
+using App.Database;
 using App.Database.EntityFramework;
 using App.ExerciseLogs;
 using App.ExerciseLogs.Models;
 using Microsoft.Extensions.DependencyInjection;
-using App.UserInterface;
 using Spectre.Console;
 
 
@@ -33,9 +33,9 @@ public class Program
         serviceCollection.AddDbContext<ExercisesDbContext>();
         serviceCollection.AddTransient(typeof(IRepositoryBase<>), typeof(RepositoryBase<>));
         serviceCollection.AddTransient<IRepositoryBase<ExerciseLog>, RepositoryBase<ExerciseLog>>();
-        serviceCollection.AddTransient(typeof(ExerciseLogsService));
+        serviceCollection.AddTransient(typeof(ExerciseService));
         serviceCollection.AddTransient(typeof(AppController));
-        serviceCollection.AddTransient(typeof(LogsController));
+        serviceCollection.AddTransient(typeof(ExerciseController));
 
         return serviceCollection.BuildServiceProvider(); ;
     }
